+++ lib/ticket_server_adapters/ticket_server_adapter.rb
@@ -1,0 +1,8 @@
+class TicketServerAdapter
+
+  def initialize(server)
+    @server = server
+    @logger = Rails.logger
+  end
+
+end
